rm unused sharp code
